Typos in the getting started tutorial (#564)
* Update getting-started.mdx

Removed extra "for" word not fitting in the sentence.

* Update using-dev-envs.mdx

Missing comma to clarify the meaning of the sentence

* Update using-dev-envs.mdx

More clarity on what is the issue

* Update using-dev-envs.mdx

There's only one endpoint created so I find that using the plural is a bit confusing.

* Update using-dev-envs.mdx

The browser is not automatically reloaded.

* Update using-dev-envs.mdx

There's only one endpoint in this example. Using the plural is confusing.

* Update using-dev-envs.mdx

It's not now that the application can be viewed. Endpoint was already created and user was able to view it before selecting the microservice.

* Update install-okteto-cli.mdx

Extra "for" removed

* Update deploy-first-env.mdx

These changes are the same already proposed in the versioned_docs